Fix click timer not being used to cancel when holding the click for too long
It was intented that the tool dropdown only shows if the user performs a regular short-click. If the mouse button is held down to, e.g. drag the SceneView camera, the context popup should be cancelled. This commit fixes the issue that the click timer was not correctly used after refactoring.