Make abstract UIE_PopupWindowContent class concrete to allow it to be serializable
The plan is to make the entire popup survive a domain reload. To make this work, we require multiple steps, the first being that the uielements content class is marked as serializable and the uielements tree is rebuild in OnEnable. This commit, however, will not yet correctly work, since the content needs to also handle serialization and likely needs to lose its legacy implementation from IMGUI times.