Reorder initialization to avoid OnGUI exception due to use of GUI.skin
On domain reload, GUI.skin will throw an exception when called from OnEnable. Relying on legacy gui for width calculation may not be ideal, but for lack of a better alternative it must suffice for now. Since gui code can only be called from within OnGUI, we are only allowed to use it when the popup first opens. However, simply omitting this step during rebuild is possible because the cached button width values are serialized anyway.